requirements.txt
streamlit==1.45.1
pandas==2.2.3
scikit-learn==1.6.1
matplotlib>=3.10.3
seaborn==0.13.2
ucimlrepo==0.0.7
streamlit_reveal_slides==0.2.0